Rename config to conf
