+++ Procfile
@@ -1,1 +1,1 @@
-web: python manage.py distributed_migrate --noinput && gunicorn config.wsgi --bind 0.0.0.0:$PORT
+web: python manage.py distributed_migrate --noinput && gunicorn conf.wsgi --bind 0.0.0.0:$PORT
