+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-DJANGO_SETTINGS_MODULE=config.settings
+DJANGO_SETTINGS_MODULE=conf.settings
